fix(types): resolve all remaining 84 mypy errors across src/ (228→0 total)
Non-core modules fixed (19 files):
- api_server.py (22): Mock metrics type ignores, CORS middleware, settings casts
- slide_generator.py (12): Optional fields, None safety guards for Path/dict
- video_composer.py (11): ProcessedSlide union type, variable annotations, casts
- test_manager.py (7): Dict value type casts for counters
- metadata_generator.py (6): str() casts for return values, tuple type fix
- source_collector.py (6): isinstance checks for BeautifulSoup Tag/NavigableString
- content_splitter.py (5): bool/str casts, type annotations, SplitContent args
- tts_integration.py (3): Dict type annotation, return value casts
- pipeline_manager.py (3): dict/list casts for return values
- 9 single-error files: type guards, casts, Optional annotations

Also: exclude api_spec_design.py from mypy (design-only file)

Mypy src/: 228 → 0 errors (76 files checked) | Tests: 107 passed

+++ src/video_editor/effect_processor.py
@@ -318,7 +318,7 @@ class EffectProcessor:
             float: イージング適用済み進行度
         """
         if easing_type == "ease_in_out":
-            return 0.5 * (1 - np.cos(progress * np.pi))
+            return float(0.5 * (1 - np.cos(progress * np.pi)))
         elif easing_type == "ease_in":
             return progress * progress
         elif easing_type == "ease_out":
